+++ Source/RimTestRedux/tests/assertion/Equals.cs
@@ -7,16 +7,6 @@ internal static class Equals
     public static void PassWhenEqual() => Assertion.Assert(1).To.Be.EqualTo(1);
 
     [Test]
-    public static void ThrowWhenNotEqual()
-    {
-        try
-        {
-            Assertion.Assert(1).To.Be.EqualTo(2);
-        }
-        catch (Exception)
-        {
-            return;
-        }
-        throw new ShouldHaveThrownException("Should have thrown an exception.");
-    }
+    [ShouldThrow]
+    public static void ThrowWhenNotEqual() => Assertion.Assert(1).To.Be.EqualTo(2);
 }