+++ Source/RimTestRedux/tests/assertion/GreaterThan.cs
@@ -7,16 +7,6 @@ internal static class GreaterThan
     public static void PassWhenGreater() => Assertion.Assert(1).To.Be.GreaterThan(0);
 
     [Test]
-    public static void ThrowWhenNotGreater()
-    {
-        try
-        {
-            Assertion.Assert(1).To.Be.GreaterThan(1);
-        }
-        catch (Exception)
-        {
-            return;
-        }
-        throw new ShouldHaveThrownException("Should have thrown an exception.");
-    }
+    [ShouldThrow]
+    public static void ThrowWhenNotGreater() => Assertion.Assert(1).To.Be.GreaterThan(1);
 }