+++ Source/RimTestRedux/tests/assertion/LessThan.cs
@@ -7,16 +7,6 @@ internal static class LessThan
     public static void PassWhenLess() => Assertion.Assert(1).To.Be.LessThan(2);
 
     [Test]
-    public static void ThrowWhenNotLess()
-    {
-        try
-        {
-            Assertion.Assert(1).To.Be.LessThan(1);
-        }
-        catch (Exception)
-        {
-            return;
-        }
-        throw new ShouldHaveThrownException("Should have thrown an exception.");
-    }
+    [ShouldThrow]
+    public static void ThrowWhenNotLess() => Assertion.Assert(1).To.Be.LessThan(1);
 }