+++ Source/RimTestRedux/tests/assertion/Null.cs
@@ -7,16 +7,6 @@ internal static class Null
     public static void PassWhenNull() => Assertion.Assert(null).To.Be.Null();
 
     [Test]
-    public static void ThrowWhenNotNull()
-    {
-        try
-        {
-            Assertion.Assert(1).To.Be.Null();
-        }
-        catch (Exception)
-        {
-            return;
-        }
-        throw new ShouldHaveThrownException("Should have thrown an exception.");
-    }
+    [ShouldThrow]
+    public static void ThrowWhenNotNull() => Assertion.Assert(1).To.Be.Null();
 }