+++ Source/RimTestRedux/tests/assertion/True.cs
@@ -7,16 +7,6 @@ internal static class True
     public static void PassWhenTrue() => Assertion.Assert(true).To.Be.True();
 
     [Test]
-    public static void ThrowWhenNotTrue()
-    {
-        try
-        {
-            Assertion.Assert(false).To.Be.True();
-        }
-        catch (Exception)
-        {
-            return;
-        }
-        throw new ShouldHaveThrownException("Should have thrown an exception.");
-    }
+    [ShouldThrow]
+    public static void ThrowWhenNotTrue() => Assertion.Assert(false).To.Be.True();
 }